Add corrections from code review
