+++ src/main/java/salesforce/ui/pages/legalentity/NewLegalEntityClassicVersion.java
@@ -1,3 +1,11 @@
+/**
+ * Copyright (c) 2021 Fundacion Jala.
+ * This software is the confidential and proprietary information of Fundacion Jala
+ * ("Confidential Information"). You shall not disclose such Confidential
+ * Information and shall use it only in accordance with the terms of the
+ * license agreement you entered into with Fundacion Jala
+ */
+
 package salesforce.ui.pages.legalentity;
 
 import core.selenium.WebElementAction;
@@ -6,7 +14,6 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import salesforce.entities.LegalEntity;
 import salesforce.ui.pages.AppPageFactory;
-
 import java.util.HashMap;
 
 import static salesforce.utils.FileTranslator.translateValue;
