+++ src/main/java/salesforce/ui/pages/legalentity/NewLegalEntityPageAbstract.java
@@ -1,8 +1,15 @@
+/**
+ * Copyright (c) 2021 Fundacion Jala.
+ * This software is the confidential and proprietary information of Fundacion Jala
+ * ("Confidential Information"). You shall not disclose such Confidential
+ * Information and shall use it only in accordance with the terms of the
+ * license agreement you entered into with Fundacion Jala
+ */
+
 package salesforce.ui.pages.legalentity;
 
 import salesforce.entities.LegalEntity;
 import salesforce.ui.pages.BasePage;
-
 import java.util.HashMap;
 import java.util.Set;
